+++ Source/PlanningExtended 1.6/Source/Definitions/PlanFilePaths.cs
@@ -1,0 +1,9 @@
+﻿using System.IO;
+using Verse;
+
+namespace PlanningExtended;
+
+internal class PlanFilePaths
+{
+    public static string PlanFolderPath => Path.Combine(GenFilePaths.SaveDataFolderPath, "Planning");
+}