+++ Source/PlanningExtended 1.6/Source/Definitions/StartupPlanVisibility.cs
@@ -1,0 +1,10 @@
+﻿namespace PlanningExtended;
+
+public enum StartupPlanVisibility
+{
+    Visible,
+
+    Invisible,
+
+    LastSaved,
+}