+++ Source/PlanningExtended 1.6/Source/Designations/PlanDesignation.cs
@@ -86,6 +86,9 @@ namespace PlanningExtended.Designations
 
         public override void DesignationDraw()
         {
+            if (_planType == PlanDesignationType.Unknown)
+                _planType = DesignationDefUtilities.GetType(def);
+
             if (!PlanAppearanceManager.IsPlanVisible(_planType))
                 return;